don't use longs unnecessarily, more finalisation, remove dead code

+++ app/src/main/java/com/greenaddress/greenapi/INotificationHandler.java
@@ -2,9 +2,9 @@ package com.greenaddress.greenapi;
 
 
 public interface INotificationHandler {
-    void onNewBlock(final long count);
+    void onNewBlock(final int count);
 
-    void onNewTransaction(final long wallet_id, final long[] subaccounts, final long value, final String txhash);
+    void onNewTransaction(final int wallet_id, final int[] subaccounts, final long value, final String txhash);
 
     void onConnectionClosed(final int code);
 }
